use Longs for stoting generic integer values

+++ community/logicaldoc/trunk/logicaldoc-core/src/main/java/com/logicaldoc/core/stats/StatsCollector.java
@@ -331,9 +331,9 @@ public class StatsCollector extends Task {
 		if (val instanceof String)
 			gen.setString1((String) val);
 		else if (val instanceof Long)
-			gen.setInteger1(((Long) val).intValue());
+			gen.setInteger1((Long) val);
 		else
-			gen.setInteger1((Integer) val);
+			gen.setInteger1(((Integer) val).longValue());
 		genericDAO.store(gen);
 	}
